Fixes and updates
