+++ fastapi_app/database/config.py
@@ -4,7 +4,7 @@ from pydantic_settings import BaseSettings
 class Settings(BaseSettings):
     """Настройки БД."""
 
-    db_url: str = 'sqlite+aiosqlite:///./db.sqlite3'
+    db_url: str = 'sqlite+aiosqlite:///database/db.sqlite3'
     db_echo: bool = True
 
 settings = Settings()
